backend/village-profile/requirements.txt
asgiref==3.4.1
bikram==1.0.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.2.6
django-chartjs==1.5.0
django-cors-headers==3.8.0
django-import-export==2.0
django-reset-migrations==0.3.1
django-utils-six==2.0
environ==1.0
envs==1.3
et-xmlfile==1.0.1
gunicorn==20.1.0
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.1
odfpy==1.4.0
openpyxl==2.5.1
pdfkit==0.6.1
Pillow==6.0.0
PyMySQL==0.9.3
python-dateutil==2.7.5
pytz==2018.7
PyYAML==5.3
six==1.12.0
sqlparse==0.4.1
tablib==0.14.0
tqdm==4.41.1
typing-extensions==3.10.0.0
Unipath==1.1
whitenoise==5.3.0
xlrd==1.2.0
xlwt==1.3.0
